requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
pillow>=9.0
xgboost>=2.0
lightgbm>=4.0
shap>=0.44
optuna>=3.0
streamlit>=1.30
plotly>=5.18
pyyaml>=6.0
python-dotenv>=1.0
joblib>=1.3
